Move Conditions to a separate source file.
